move sdp utils to separate module
